+++ src/operating_systems/freebsd/qemu_vm.ts
@@ -5,10 +5,10 @@ export class QemuVm extends Vm {
     // prettier-ignore
     return [
       '-device', 'virtio-blk-pci,drive=drive0,bootindex=0',
-      '-drive', `if=none,file=${this.configuration.diskImage},id=drive0,cache=writeback,discard=ignore,format=raw`,
+      '-drive', `if=none,file=${this.configuration.diskImage},id=drive0,cache=unsafe,discard=ignore,format=raw`,
 
       '-device', 'virtio-blk-pci,drive=drive1,bootindex=1',
-      '-drive', `if=none,file=${this.configuration.resourcesDiskImage},id=drive1,cache=writeback,discard=ignore,format=raw`,
+      '-drive', `if=none,file=${this.configuration.resourcesDiskImage},id=drive1,cache=unsafe,discard=ignore,format=raw`,
     ]
   }
 }