+++ src/qemu_vm.ts
@@ -61,10 +61,10 @@ export abstract class Vm extends vm.Vm {
       '-device', 'virtio-scsi-pci',
 
       '-device', 'scsi-hd,drive=drive0,bootindex=0',
-      '-drive', `if=none,file=${this.configuration.diskImage},id=drive0,cache=writeback,discard=ignore,format=raw`,
+      '-drive', `if=none,file=${this.configuration.diskImage},id=drive0,cache=unsafe,discard=ignore,format=raw`,
 
       '-device', 'scsi-hd,drive=drive1,bootindex=1',
-      '-drive', `if=none,file=${this.configuration.resourcesDiskImage},id=drive1,cache=writeback,discard=ignore,format=raw`,
+      '-drive', `if=none,file=${this.configuration.resourcesDiskImage},id=drive1,cache=unsafe,discard=ignore,format=raw`,
     ]
   }